new dir structure added simple_node example added more methods according to protocol grammar
